+++ components/commercetools-ui/organisms/checkout/ct-checkout/index.tsx
@@ -1,18 +1,16 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useParams } from 'next/navigation';
-//import { checkout } from '@commercetools/checkout-browser-sdk';
-//import { getLocalizationInfo } from 'project.config';
-import { useAccount, useCart, useProjectSettings } from 'frontastic';
+import { checkout, init } from '@commercetools/checkout-browser-sdk';
+import { useCart, useProjectSettings, useCheckoutToken } from 'frontastic';
 import { CheckoutWrappedProps } from '..';
 import Header from '../components/header';
 
 const CommercetoolsCheckout = ({ logo, ...emptyState }: CheckoutWrappedProps) => {
   const initiatedCheckout = useRef(false);
 
-  const { data: cart } = useCart();
+  const [isLoading, setIsLoading] = useState(true);
 
-  const { account } = useAccount();
-  console.log(account);
+  const { data: cart } = useCart();
 
   const { data: projectSettings } = useProjectSettings();
 
@@ -20,30 +18,63 @@ const CommercetoolsCheckout = ({ logo, ...emptyState }: CheckoutWrappedProps) =>
 
   const { projectKey } = projectSettings ?? {};
 
+  const accessToken = useCheckoutToken();
+
+  const applicationId = useMemo(() => {
+    return locale === 'en'
+      ? process.env.NEXT_PUBLIC_COMMERCETOOLS_CHECKOUT_APPLICATION_ID_EN
+      : process.env.NEXT_PUBLIC_COMMERCETOOLS_CHECKOUT_APPLICATION_ID_DE;
+  }, [locale]);
+
   useEffect(() => {
-    if (initiatedCheckout.current || !projectKey || !cart) return;
+    if (initiatedCheckout.current || !projectKey || !cart || !accessToken) return;
 
     initiatedCheckout.current = true;
 
-    //const info = getLocalizationInfo(locale);
+    const host = typeof window !== 'undefined' ? window.location.origin : '';
 
-    // checkout({
-    //   host: process.env.NEXT_PUBLIC_COMMERCETOOLS_CHECKOUT_HOST,
-    //   sellerId: projectKey,
-    //   applicationId: process.env.NEXT_PUBLIC_COMMERCETOOLS_CHECKOUT_APPLICATION_ID,
-    //   callbackUrl: process.env.NEXT_PUBLIC_COMMERCETOOLS_CHECKOUT_CALLBACK_URL,
-    //   cartId: cart.cartId,
-    //   accessToken: '{accessToken}',
-    //   locale: info.locale,
-    // });
-  }, [projectKey, cart, locale]);
+    init({
+      checkoutConfig: {},
+      onInfo: (message) => {
+        if (message.code === 'checkout_loaded') setIsLoading(false);
+      },
+    });
+
+    checkout({
+      host: process.env.NEXT_PUBLIC_COMMERCETOOLS_CHECKOUT_HOST,
+      sellerId: projectKey,
+      applicationId,
+      callbackUrl: `${host}/${process.env.NEXT_PUBLIC_COMMERCETOOLS_CHECKOUT_CALLBACK_URL}`,
+      cartId: cart.cartId,
+      accessToken,
+      locale,
+    });
+  }, [projectKey, cart, locale, accessToken, applicationId]);
+
+  const handleGoingBack = useCallback(() => {
+    const checkoutIframe = document.getElementById('ctc-wrapper');
+
+    if (checkoutIframe) checkoutIframe.remove();
+
+    window.onpopstate = null;
+  }, []);
+
+  useEffect(() => {
+    window.onpopstate = handleGoingBack;
+  }, [handleGoingBack]);
 
   return (
     <div className="min-h-screen lg:bg-neutral-200">
       <Header logo={logo} {...emptyState} />
       <div className="relative">
-        <div data-ctc />
+        {/* eslint-disable-next-line tailwindcss/no-custom-classname*/}
+        <div data-ctc className="checkout-Container" />
       </div>
+      {isLoading && (
+        <div className="flex h-[80vh] items-center justify-center">
+          <div className="loading-full-screen" />
+        </div>
+      )}
     </div>
   );
 };