+++ frontastic/hooks/useCart/types.ts
@@ -36,6 +36,7 @@ export interface UseCartReturn {
   shippingMethods: { data?: ShippingMethod[] };
   orderCart: () => Promise<Order>;
   getOrder: (orderId: string) => Promise<Order>;
+  getCheckoutOrder: (orderId: string) => Promise<Order>;
   orderHistory?: () => Promise<Order[]>;
   getProjectSettings?: () => Promise<ProjectSettings>;
 }